+++ opal/models.py
@@ -771,13 +771,7 @@ class Episode(UpdatesFromDictMixin, TrackedModel):
         Validates that the stage being set is appropriate for the category
         and raises ValueError if not.
         """
-        if not self.category.has_stage(stage):
-            if stage is not None:
-                msg = "Can't set stage to {0} for {1} Episode".format(
-                    stage, self.category.display_name
-                )
-                raise ValueError(msg)
-        self.stage = stage
+        self.category.set_stage(stage, user, data)
 
     def set_tag_names(self, tag_names, user):
         """