+++ CHANGELOG.md
@@ -24,7 +24,7 @@ When you delete a user, it will no longer delete all related episodes and subrec
 
 Episode categories now enforce a set of valid `Episode.stage` values.
 `EpisodeCategory` now includes the `.get_stages()` and `.has_stage(stage)` methods,
-while `Episode` has a `set_stage` setter which is used by the JSON API.
+while `Episode` has a `set_stage` setter which is used by the UpdateFromDictMixin JSON API.
 
 #### lookuplists.lookuplists
 
@@ -85,6 +85,13 @@ failure.
 
 If you are a plugin developer upgrading an existing plugin you will have to
 manually add support for `--failfast` passthrough to your `runtests.py`.
+
+
+#### Moves scaffold to be a django management command
+
+The rest of the api is still the same but now
+we run `python manage.py scaffold {my_app_name}`
+
 
 #### Misc Changes